Added the following configuration options to the configuration file:
- AddDates
FALSE: no behavior changes
TRUE: The download date will be put in the front of the filename of the directory name (if CreateFolder is true). The program tries to be smart enough to recognizes identical files, even if you keep changing the setting for 'AddDates' between runs

- UserAgentIds
- UserAgentStrings
Some download sites want the UserAgent string set to a specific value to allow download. The UserAgentIDs should include a comma delimited list of web site indentifications (unique parts in the download url for the trailer). UserAgentStrings is the matching comma-delimited list of user agent strings which will be used if a download from a UserAgentId-site is happening. 
for example:
 
<add key="UserAgentIds" value="apple.com" />
<add key="UserAgentStrings" value="QuickTime/7.6.2" />
 
If the trailer is downloaded using a URL containing "apple.com" the user agent on the http object will be set to "QuickTime/7.6.2"

- EmailReturnAddress
if status email is sent from the program, this is the return address in the status email

- EmailReturnDisplayName
the display name of the return address in the status mail